+++ CHANGELOG.rst
@@ -12,6 +12,9 @@ To be released
 * Added ``client.relations.block`` and ``client.relations.unblock`` for blocking/unblocking users.
 * Implemented ``/api/games/export/imports`` under
   ``client.games.export_imported``.
+* Implemented ``/api/study/by/{username}`` under
+  ``client.studies.list_user_studes`` to get the metadata of all studies
+  of a given user.
 
 Thanks to all the contributors who helped to this release:
 - @hsheth2